requirements.txt
cycler==0.10.0
graphviz==0.13
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
Pillow==6.2.1
pyparsing==2.4.2
python-dateutil==2.8.0
six==1.12.0
torch==1.3.0.post2
torchvision==0.4.1.post2
torchviz==0.0.1
